requirements.txt
asgiref==3.3.4
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.7
daphne==3.0.2
Django==3.2
django-cors-headers==3.7.0
django-rest==0.8.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
docutils==0.17.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
Flask==1.1.2
hyperlink==21.0.0
idna==2.10
ifaddr==0.1.7
importlib-metadata==4.0.1
incremental==21.3.0
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
keyring==23.0.1
lml==0.1.0
MarkupSafe==1.1.1
MouseInfo==0.1.3
# mysqlclient @ file:///E:/demo%20projcts/algotarianbackend/mysqlclient-1.4.6-cp37-cp37m-win32.whl
Naked==0.1.31
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.2.0
pkginfo==1.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
# PyAudio @ file:///E:/PyAudio-0.2.11-cp37-cp37m-win32.whl
pyauto==0.0.2
PyAutoGUI==0.9.52
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
# pycryptosat @ file:///E:/python%20sabpaisa/pycryptosat-0.6.1.1-cp37-cp37m-win32.whl
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyGetWindow==0.0.9
Pygments==2.8.1
PyJWT==2.1.0
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
selenium==3.141.0
service-identity==18.1.0
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tqdm==4.60.0
twine==3.4.1
Twisted==21.2.0
# twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.4.0
